Change security strucure, add adnotations in security classes caused spring profile issues. CHECKPOINT before refactor security from package into modules
