+++ gaminatorium-main/monolith/src/main/java/eu/gaminatorium/security/managers/GaminatoriumAuthenticationManager.java
@@ -2,6 +2,7 @@ package eu.gaminatorium.security.managers;
 
 import eu.gaminatorium.security.providers.GaminatoriumAuthenticationProvider;
 import lombok.AllArgsConstructor;
+import org.springframework.context.annotation.Profile;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
@@ -9,6 +10,7 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Component;
 
 @Component
+@Profile(value = "prod")
 @AllArgsConstructor
 public class GaminatoriumAuthenticationManager implements AuthenticationManager {
